alltray: Bring in initial support for alltray
alltray allows Spotify to be left running in the Linux system tray.
Current versions of alltray do not work well with modern desktops but by
pulling in an older but still compatible version of libgtop-2.0 a previous
version of alltray will work fine.

The previous method for launching spotify inside the container did not
play nicely with alltray, however, so at the same time we'll eliminate the
launcher script, make the container always run as an unprivileged user
(rather than exec-ing to it in the launcher script) and select sensible
defaults for certain environment variables along the way.  This does
streamline the container startup somewhat at the potential cost of some
flexibility.

Signed-off-by: Joe MacDonald <joe_macdonald@mentor.com>